pyproject.toml
[tool.poetry]
name = "my-first-project"
version = "0.1.0"
description = "A simple project setup"
authors = ["Your Name <your-email@example.com>"]


# Добавьте необходимые зависимости (если они есть)
[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.0"

[tool.poetry.dev-dependencies]
ruff = "^0.11.0"

